Revert "Updated components related to entity stats to reflect complexity of current game."
This reverts commit 191348ecf860030485a4914c53cc6372958306df.

+++ source/core/src/main/com/deco2800/game/components/player/PlayerStatsDisplay.java
@@ -43,8 +43,8 @@ public class PlayerStatsDisplay extends UIComponent {
     heartImage = new Image(ServiceLocator.getResourceService().getAsset("images/heart.png", Texture.class));
 
     // Health text
-    float health = entity.getComponent(CombatStatsComponent.class).getHealth();
-    CharSequence healthText = String.format("Health: %f", health);
+    int health = entity.getComponent(CombatStatsComponent.class).getHealth();
+    CharSequence healthText = String.format("Health: %d", health);
     healthLabel = new Label(healthText, skin, "large");
 
     table.add(heartImage).size(heartSideLength).pad(5);
